requirements.txt
flask==3.1.0
gunicorn==23.0.0
opencv-python-headless==4.11.0.86
numpy==2.2.4
pillow==11.1.0
